refactor: Split out config elements
Split config sections into individual modules for the better
maintainability.
