+++ src/config/resource.rs
@@ -1,0 +1,32 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+use serde::Deserialize;
+
+use crate::config::common::default_sql_driver;
+
+/// Resource provider (domain, project).
+#[derive(Debug, Deserialize, Clone)]
+pub struct ResourceProvider {
+    /// Resource provider backend.
+    #[serde(default = "default_sql_driver")]
+    pub driver: String,
+}
+
+impl Default for ResourceProvider {
+    fn default() -> Self {
+        Self {
+            driver: default_sql_driver(),
+        }
+    }
+}
