+++ src/config/revoke.rs
@@ -1,0 +1,42 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+use serde::Deserialize;
+
+use crate::config::common::default_sql_driver;
+
+/// Revoke provider configuration.
+#[derive(Debug, Deserialize, Clone)]
+pub struct RevokeProvider {
+    /// Entry point for the token revocation backend driver in the
+    /// `keystone.revoke` namespace. Keystone only provides a `sql` driver.
+    #[serde(default = "default_sql_driver")]
+    pub driver: String,
+    /// The number of seconds after a token has expired before a corresponding
+    /// revocation event may be purged from the backend.
+    #[serde(default = "default_expiration_buffer")]
+    pub expiration_buffer: usize,
+}
+
+impl Default for RevokeProvider {
+    fn default() -> Self {
+        Self {
+            driver: default_sql_driver(),
+            expiration_buffer: default_expiration_buffer(),
+        }
+    }
+}
+
+fn default_expiration_buffer() -> usize {
+    1800
+}
